[Q20-22] Added methods for displaying, resolved some displaying errors,some are yet to be resolved

+++ quadriga/src/main/java/edu/asu/diging/quadriga/web/AssignAppsController.java
@@ -2,27 +2,30 @@ package edu.asu.diging.quadriga.web;
 
 import java.util.HashSet;
 
-import java.util.Iterator;
+
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
 import org.springframework.beans.factory.annotation.Autowired;
+
+
 import org.springframework.security.access.prepost.PreAuthorize;
+
 import org.springframework.security.web.csrf.CsrfToken;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import edu.asu.diging.quadriga.core.model.users.SimpleUserApp;
 import edu.asu.diging.quadriga.core.service.SimpleUserAppService;
 
 @Controller
 public class AssignAppsController {
-    private final Logger logger = LoggerFactory.getLogger(getClass());
+    
 
     @Autowired
     private SimpleUserAppService simpleUserAppService;
@@ -35,20 +38,28 @@ public class AssignAppsController {
 
         model.addAttribute("username", username);
         model.addAttribute("userApps", userApps);
-        model.addAttribute("csrfToken",csrfToken);
-        System.out.println(csrfToken);
+        model.addAttribute("csrfToken",csrfToken.getToken());
         return "admin/user/apps";
     }
     @PreAuthorize("isAuthenticated()")
     @RequestMapping(value = "/admin/user/{username}/app/{clientId}/assign", method = RequestMethod.POST)
-    public String assign(@PathVariable String username, @PathVariable String clientId) {
-        
+    public String assign(@PathVariable String username, @PathVariable String clientId,@RequestParam("_csrf") String csrfToken,HttpServletRequest request){
+        CsrfToken storedCsrf = (CsrfToken) request.getAttribute(CsrfToken.class.getName());
+        if(!storedCsrf.getToken().equals(csrfToken))
+        {
+            return "Not Authorized";
+        }
         simpleUserAppService.save(username, clientId);
         return "redirect: /quadriga/admin/user/" + username + "/apps";
     }
     @PreAuthorize("isAuthenticated()")
     @RequestMapping(value = "/admin/user/{username}/app/{clientId}/revoke", method = RequestMethod.POST)
-    public String withdraw(@PathVariable String username, @PathVariable String clientId) {
+    public String withdraw(@PathVariable String username, @PathVariable String clientId,@RequestParam("_csrf") String csrfToken,HttpServletRequest request ) {
+        CsrfToken storedCsrf = (CsrfToken) request.getAttribute(CsrfToken.class.getName());
+        if(!storedCsrf.getToken().equals(csrfToken))
+        {
+            return "Not Authorized";
+        }
         
         simpleUserAppService.delete(username, clientId);
         return "redirect: /quadriga/admin/user/" + username + "/apps";
